+++ NEWS.md
@@ -7,9 +7,10 @@ The current version of the DLMtool package is available for download from [CRAN]
 - add option to specify width of the CAL bins using `OM@cpars$binWidth`
 - change MGT calculation from Z to M
 
+
 ### Fixes
 - fix indexing issue for recruitment deviations
-
+- fix issue with `cpars$mov` when nareas>2